fix: rename problem-sloved-chart.jsx to problem-solved-chart.jsx
